src/api-app/requirements.txt
pillow
httpx
python-dotenv
fastapi
langchain_openai
streamlit